do-core1: Move instruction unit tests to do-core
And define a workspace.
