+++ Cargo.toml
@@ -7,3 +7,8 @@ edition = "2018"
 [dependencies]
 clap = "2.33.3"
 do-core = { path = "do-core" }
+
+[workspace]
+members = [
+    "do-core"
+]
